requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.25
pymysql==1.1.0
cryptography>=42.0.0
reportlab==4.1.0
pypdf==4.0.1
arabic-reshaper==3.0.0
python-bidi>=0.4.2
python-dotenv==1.0.0
Pillow>=10.0.0
hijri-converter>=2.3.1
